requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

biopython==1.84
colorama==0.4.6
einops==0.8.0
matplotlib==3.10.0
more-itertools==10.5.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
PyYAML==6.0.2
rotary-embedding-torch==0.8.6
scikit-learn==1.6.0
scipy==1.14.1
tabulate==0.9.0
timm==1.0.12
tqdm==4.67.1
wandb==0.19.1
xformers==0.0.29.post1
